pyproject.toml
[project]
name = "yacb"
version = "0.1.0"
description = "Personal AI assistant daemon - Telegram, WhatsApp, Discord"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-telegram-bot[socks]>=21.0",
    "discord.py>=2.3.0",
    "httpx[socks]>=0.25.0",
    "neonize>=0.3.14",
    "croniter>=2.0.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.19.0",
    "readability-lxml>=0.8.0",
    "tavily-python>=0.7.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
yacb = "core.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
